[ion] Simulator: haptics feedback are only implemented on Android and depends on the System settings
